chore: correcting the structure
