+++ src/components/utils/checkConfiguration.ts
@@ -1,44 +1,4 @@
-import type { Configuration, PaymentMethodsResponse } from '../core';
-
-export const checkPaymentMethodsResponse = (
-  paymentMethodsResponse: PaymentMethodsResponse | undefined
-) => {
-  if (paymentMethodsResponse === undefined) {
-    throw new Error(
-      'paymentMethodsResponse is undefined. Make sure to make POST `paymentMethods` or call `conext.startSession()`' +
-        'Try JSON.parse("{...}") your paymentMethodsResponse.'
-    );
-  }
-
-  if (typeof paymentMethodsResponse === 'string') {
-    throw new TypeError(
-      'paymentMethodsResponse was provided but of an incorrect type (should be an object but a string was provided).' +
-        'Try JSON.parse("{...}") your paymentMethodsResponse.'
-    );
-  }
-
-  if (Array.isArray(paymentMethodsResponse)) {
-    throw new TypeError(
-      'paymentMethodsResponse was provided but of an incorrect type (should be an object but an array was provided).' +
-        'Please check you are passing the whole response.'
-    );
-  }
-
-  if (
-    !paymentMethodsResponse?.paymentMethods?.length &&
-    !paymentMethodsResponse?.storedPaymentMethods?.length
-  ) {
-    console.warn(
-      'paymentMethodsResponse was provided but no payment methods were found.'
-    );
-  }
-
-  return paymentMethodsResponse;
-};
-
-const countryCodeRegex = /^[A-Z]{2}$/;
-const currencyCodeRegex = /^[A-Z]{3}$/;
-const clientKeyRegex = /^[a-z]{4,8}_[a-zA-Z0-9]{8,128}$/;
+import type { Configuration } from '../../core';
 
 export const checkConfiguration = (configuration: Configuration) => {
   if (!configuration?.returnUrl) {
@@ -86,3 +46,7 @@ export const checkConfiguration = (configuration: Configuration) => {
     );
   }
 };
+
+const countryCodeRegex = /^[A-Z]{2}$/;
+const currencyCodeRegex = /^[A-Z]{3}$/;
+const clientKeyRegex = /^[a-z]{4,8}_[a-zA-Z0-9]{8,128}$/;
